+++ packages/bs4/src/binders/index.ts
@@ -3,5 +3,6 @@ export { collapseBinder } from './bs4-collapse.binder';
 export { dropdownBinder } from './bs4-dropdown.binder';
 export { expanOnUrlBinder } from './bs4-expan-on-url.binder';
 export { scrollspyClassBinder } from './bs4-scrollspy-class.binder';
+export { toggleItemBinder } from './bs4-toggle-item.binder';
 export { tooltipBinder } from './bs4-tooltip.binder';
 export { scrollToOnEventBinder } from './scroll-to-on-event.binder';