+++ packages/bs4/src/constants.ts
@@ -1,0 +1,17 @@
+// EventDispatcher events
+export const TOGGLE_BUTTON = {
+  nsPrefix: 'bs4-toggle-button:',
+  eventNames: {
+    toggle: 'toggle',
+    toggled: 'toggled',
+    init: 'init',
+    state: 'state',
+  }
+};
+
+export const TOGGLE_ITEM = {
+  elEventNames: {
+    show: 'show',
+    hide: 'hide',
+  }
+}