+++ packages/bs4/src/index.ts
@@ -4,4 +4,5 @@ export * from './components';
 // export * as formatters from './formatters/bs4.formatters';
 export * from './services';
 // export * from './interfaces/interfaces';
+export * from './constants';
 export * from './bs4.module';